+++ VMThirdPartyTransfer.cs
@@ -1,6 +1,8 @@
+using System;
+
 public class VMThirdPartyTransfer{
     public decimal TransferAmount{get;set;}
-    public int RecipientBankAccountNumber { get; set; }
+    public Int64 RecipientBankAccountNumber { get; set; }
 
     public string RecipientBankAccountName { get; set; }
 }
